常用redis数据结构

+++ src/main/resources/init.sql
@@ -10,7 +10,7 @@ SET FOREIGN_KEY_CHECKS = 0;
 DROP TABLE IF EXISTS `GOODS`;
 create table GOODS
 (
-    ID         decimal(8) primary key  auto_increment,
+    ID         int primary key  auto_increment,
     NAME       varchar(50),
     GOOD_ID    decimal(8),
     LEFT_COUNT decimal(8)
